requirements.txt
# all_new_cbot Python dependencies
python-telegram-bot==22.6
python-dotenv==1.1.1
beautifulsoup4>=4.12
# Install browser once: python -m playwright install chromium
playwright>=1.40
Pillow>=10.0
requests>=2.31
pytest>=8.0
